lf iterate: lfops: consolidate pr and landing commands from lfpr
Merged `lfpr` CLI into `lfops` as subcommands. Commands are now `lfops pr create`, `lfops land`, and `lfops commit`. Also updated documentation and removed the standalone `lfpr` entry point.
