+++ pyproject.toml
@@ -36,7 +36,6 @@ dev = [
 lf = "loopflow.cli:main"
 lfd = "loopflow.lfd:main"
 lfwt = "loopflow.lfwt:main"
-lfpr = "loopflow.lfpr:main"
 lfops = "loopflow.lfops:main"
 
 [build-system]
